Figure9 10 (#205)
* kl divergence function

* KL for bern and gamma

* KL

* notebook updated

* some more updates to the KL

* the KL function added to analyze

* figure 12 added Kl

* linting with autopep8

* oops!

* notebook and add more points to KL

* quick fix

* clean up and address xomments

* moved the figures to figure12.py

* some adjustments

* a slight change to the figure 12

* figure 12 more changes

* figure 12 finally!

* final commit

* linting

* linting

* lint

* linting

* linting

* lint

* lint

* linting

* lint

* lint final

* last linting

* linting done

* kl removed and loop added

* small error fixed

* figure 2

* size=1500 observation

* warning fixed

* warnings remoooved

* reduced the number of iters

* warning for analyze

* more warnings

* some more warnigns

* figure 9 and 10 combined

* figures 7 and 8 combined

* deleting figure7

* autopep8 action fixes (#201)

* missing line

* more work:

* combine functions inside

* moving average

* autopep8 action fixes (#203)

* doc strings

* reorganizing the figures

* autopep8 action fixes (#204)

* linting

* lint
