+++ lineage/figures/figure3.py
@@ -2,8 +2,6 @@
 File: figure3.py
 Purpose: Generates figure 3.
 """
-import numpy as np
-import matplotlib.pyplot as plt
 
 from .figureCommon import getSetup
 
@@ -13,6 +11,6 @@ def makeFigure():
     Makes figure 3.
     """
 
-    ax, f = getSetup((21, 6), (1, 3))
+    _, f = getSetup((21, 6), (1, 3))
 
     return f
