+++ lineage/figures/figure6.py
@@ -3,8 +3,6 @@ File: figure6.py
 Purpose: Generates figure 6.
 
 """
-import numpy as np
-import matplotlib.pyplot as plt
 
 from .figureCommon import getSetup
 
@@ -14,6 +12,6 @@ def makeFigure():
     Makes figure 6.
     """
 
-    ax, f = getSetup((21, 6), (1, 3))
+    _, f = getSetup((21, 6), (1, 3))
 
     return f
